--- main.go
@@ -1,244 +1,0 @@
-// Recipes API
-//
-// This is a sample recipes API. You can find out more about the API at https://github.com/richinex/recipes-api.
-//
-//		Schemes: http
-//	 Host: localhost:8080
-//		BasePath: /
-//		Version: 1.0.0
-//		Contact: Richard Chukwu <richinex@gmail.com>
-//
-//		Consumes:
-//		- application/json
-//
-//		Produces:
-//		- application/json
-//
-// swagger:meta
-package main
-
-import (
-	"context"
-	"fmt"
-	"log"
-	"net/http"
-	"os"
-	"strings"
-	"time"
-
-	"github.com/gin-gonic/gin"
-	"go.mongodb.org/mongo-driver/bson"
-	"go.mongodb.org/mongo-driver/bson/primitive"
-	"go.mongodb.org/mongo-driver/mongo"
-	"go.mongodb.org/mongo-driver/mongo/options"
-	"go.mongodb.org/mongo-driver/mongo/readpref"
-)
-
-var recipes []Recipe
-
-// swagger:parameters recipes newRecipe
-type Recipe struct {
-	//swagger:ignore
-	ID           primitive.ObjectID `json:"id" bson:"_id"`
-	Name         string             `json:"name" bson:"name"`
-	Tags         []string           `json:"tags" bson:"tags"`
-	Ingredients  []string           `json:"ingredients" bson:"ingredients"`
-	Instructions []string           `json:"instructions" bson:"instructions"`
-	PublishedAt  time.Time          `json:"publishedAt" bson:"publishedAt"`
-}
-
-// swagger:operation GET /recipes recipes listRecipes
-// Returns list of recipes
-// ---
-// produces:
-// - application/json
-// responses:
-//
-//	'200':
-//	    description: Successful operation
-func ListRecipesHandler(c *gin.Context) {
-	collection := client.Database(os.Getenv("MONGO_DATABASE")).Collection("recipes")
-	cur, err := collection.Find(ctx, bson.M{})
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-	defer cur.Close(ctx)
-	recipes := make([]Recipe, 0)
-	for cur.Next(ctx) {
-		var recipe Recipe
-		cur.Decode(&recipe)
-		recipes = append(recipes, recipe)
-	}
-	c.JSON(http.StatusOK, recipes)
-}
-
-// swagger:operation POST /recipes recipes newRecipe
-// Create a new recipe
-// ---
-// produces:
-// - application/json
-// responses:
-//
-//	'200':
-//	    description: Successful operation
-//	'400':
-//	    description: Invalid input
-func NewRecipeHandler(c *gin.Context) {
-	collection := client.Database(os.Getenv("MONGO_DATABASE")).Collection("recipes")
-	var recipe Recipe
-	if err := c.ShouldBindJSON(&recipe); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-	recipe.ID = primitive.NewObjectID()
-	recipe.PublishedAt = time.Now()
-	_, err = collection.InsertOne(ctx, recipe)
-	if err != nil {
-		fmt.Println(err)
-		c.JSON(http.StatusInternalServerError,
-			gin.H{"error": "Error while inserting a new recipe"})
-		return
-	}
-	c.JSON(http.StatusOK, recipe)
-
-}
-
-// swagger:operation PUT /recipes/{id} recipes updateRecipe
-// Update an existing recipe
-// ---
-// parameters:
-//   - name: id
-//     in: path
-//     description: ID of the recipe
-//     required: true
-//     type: string
-//
-// produces:
-// - application/json
-// responses:
-//
-//	'200':
-//	    description: Successful operation
-//	'400':
-//	    description: Invalid input
-//	'404':
-//	    description: Invalid recipe ID
-func UpdateRecipeHandler(c *gin.Context) {
-	collection := client.Database(os.Getenv("MONGO_DATABASE")).Collection("recipes")
-	id := c.Param("id")
-	var recipe Recipe
-	if err := c.ShouldBindJSON(&recipe); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-	objectID, _ := primitive.ObjectIDFromHex(id)
-	_, err = collection.UpdateOne(ctx, bson.M{
-		"_id": objectID,
-	}, bson.D{{"$set", bson.D{
-		{"name", recipe.Name},
-		{"instructions", recipe.Instructions},
-		{"ingredients", recipe.Ingredients},
-		{"tags", recipe.Tags},
-	}}})
-	if err != nil {
-		fmt.Println(err)
-		c.JSON(http.StatusInternalServerError,
-			gin.H{"error": "Error while updating a recipe"})
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{"message": "Recipe has been updated"})
-}
-
-// swagger:operation DELETE /recipes/{id} recipes deleteRecipe
-// Delete an existing recipe
-// ---
-// produces:
-// - application/json
-// parameters:
-//   - name: id
-//     in: path
-//     description: ID of the recipe
-//     required: true
-//     type: string
-//
-// responses:
-//
-//	'200':
-//	    description: Successful operation
-//	'404':
-//	    description: Invalid recipe ID
-func DeleteRecipeHandler(c *gin.Context) {
-	collection := client.Database(os.Getenv("MONGO_DATABASE")).Collection("recipes")
-	id := c.Param("id")
-	objectId, _ := primitive.ObjectIDFromHex(id)
-	_, err := collection.DeleteOne(ctx, bson.M{
-		"_id": objectId,
-	})
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-	c.JSON(http.StatusOK, gin.H{"message": "Recipe has been deleted"})
-}
-
-// swagger:operation GET /recipes/search recipes findRecipe
-// Search recipes based on tags
-// ---
-// produces:
-// - application/json
-// parameters:
-//   - name: tag
-//     in: query
-//     description: recipe tag
-//     required: true
-//     type: string
-//
-// responses:
-//
-//	'200':
-//	    description: Successful operation
-func SearchRecipesHandler(c *gin.Context) {
-	tag := c.Query("tag")
-	listOfRecipes := make([]Recipe, 0)
-
-	for i := 0; i < len(recipes); i++ {
-		found := false
-		for _, t := range recipes[i].Tags {
-			if strings.EqualFold(t, tag) {
-				found = true
-			}
-		}
-		if found {
-			listOfRecipes = append(listOfRecipes, recipes[i])
-		}
-
-	}
-
-	c.JSON(http.StatusOK, listOfRecipes)
-}
-
-var ctx context.Context
-var err error
-var client *mongo.Client
-
-func init() {
-	ctx = context.Background()
-	client, err = mongo.Connect(ctx, options.Client().ApplyURI(os.Getenv("MONGO_URI")))
-	if err = client.Ping(context.TODO(), readpref.Primary()); err != nil {
-		log.Fatal(err)
-	}
-	log.Println("Connected to MongoDB!")
-
-}
-
-func main() {
-	router := gin.Default()
-	router.POST("/recipes", NewRecipeHandler)
-	router.GET("/recipes", ListRecipesHandler)
-	router.PUT("/recipes/:id", UpdateRecipeHandler)
-	router.DELETE("/recipes/:id", DeleteRecipeHandler)
-	router.GET("/recipes/search", SearchRecipesHandler)
-	router.Run()
-}
